fix winit build
